fix: 프로필 모달 컴포넌트로 분리
